첨부파일 없는 게시물 삭제

+++ src/main/java/org/zerock/service/board/BoardSerivce.java
@@ -143,8 +143,10 @@ public class BoardSerivce {
 		
 		File[] listFiles = folder.listFiles();
 		
-		for (File file : listFiles) {
-			file.delete();
+		if (listFiles != null) {
+			for (File file : listFiles) {
+				file.delete();
+			}
 		}
 		
 		folder.delete();
